fix(io): fix bug when using relative pathName
XDMF files assume the h5 files they reference are in the same directory.
The full pathName of the h5 file was being written to the xmf files.
If the h5 files were moved, or if a relative pathName was used, then the
reference to the h5 files within the xmf files was no longer correct,
resulting in read errors.  Now, only the filename (not path) of the h5
file is referenced in the xmf files.