fix(mpi): re-implement ordered halo exchanges
Ordered exchanges update corner and edge boundary elements and are
required when stencils with diagonal components are used.  Currently the
laplace operators for viscosity and cequations rely on diagonal stencil
elements.  When there are no numerics that require these additional
boundary elements to be exchanged, then the unordered exchanges
developed by the mpi team are used.